api/requirements.txt
fastapi
uvicorn
sse-starlette
kafka-python
